Remove consistent-return rule
This rule causes far more problems than it's worth, and ends up mucking
with semantics of proper code. If there were a way to mark
`return undefined` as an ambiguous expression, I think this would be a
different matter. As it currently stands, this rule catches some
occasional bugs but mostly causes pain.